+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added optional dst_rsc_name to backup_schedule_enable
+
 ## [1.26.0] - 2025-08-18
 
 ### Added